feat: adicionar tabela de questionários de estilo de vida
- Criar migration V4 com tabela tbl_questionarios_estilo_vida
- Adicionar campos de rotina (objetivo, frequência e tempo de treino)
- Incluir campos de saúde (cirurgias, doenças, medicamentos,
suplementos, uso de anabolizantes)
- Adicionar campos de hábitos (fumo, álcool, sono, intestino, ingestão
de água)
- Incluir preferências alimentares (alimentos que não gosta, frutas,
número de refeições)
- Adicionar campos clínicos (pressão arterial, intolerâncias)
- Implementar relacionamento OneToOne com tbl_consultas via consulta_id
- Configurar CASCADE delete e índice para performance
- Adicionar timestamps de auditoria (created_at, updated_at)

+++ src/main/resources/db/migration/V4__create_table_questionarios_estilo_vida.sql
@@ -26,7 +26,7 @@ CREATE TABLE tbl_questionarios_estilo_vida (
     frequencia_alcool VARCHAR(255),
     funcionamento_intestino VARCHAR(255),
     qualidade_sono VARCHAR(255),
-    ingestao_agua_diaria DECIMAL(5,2),
+    ingestao_agua_diaria DOUBLE,
 
     -- Preferências
     alimentos_nao_gosta TEXT,
